비동기 action도 slice로 가져오기
