--- ToGoFurther/src/actions.js
@@ -1,103 +1,0 @@
-import {
-  fetchRegions,
-  fetchCategories,
-  fetchRestaurants,
-  fetchRestaurant,
-  postLogin,
-  postReview,
-} from './services/api';
-
-import { saveItem } from './services/storage';
-
-import { actions } from './slice';
-
-export const {
-  setRegions,
-  setCategories,
-  setRestaurant,
-  setRestaurants,
-  selectRegion,
-  selectCategory,
-  changeLoginField,
-  changeReviewField,
-  setAccessToken,
-  logout,
-  clearReviewFields,
-  setReviews,
-} = actions;
-
-export function loadInitialData() {
-  return async (dispatch) => {
-    const regions = await fetchRegions();
-    dispatch(setRegions(regions));
-
-    const categories = await fetchCategories();
-    dispatch(setCategories(categories));
-  };
-}
-
-export function loadRestaurants() {
-  return async (dispatch, getState) => {
-    const {
-      selectedRegion: region,
-      selectedCategory: category,
-    } = getState();
-
-    if (!region || !category) {
-      return;
-    }
-
-    const restaurants = await fetchRestaurants({
-      regionName: region.name,
-      categoryId: category.id,
-    });
-    dispatch(setRestaurants(restaurants));
-  };
-}
-
-export function loadRestaurant({ restaurantId }) {
-  return async (dispatch) => {
-    dispatch(setRestaurant(null));
-
-    const restaurant = await fetchRestaurant({ restaurantId });
-
-    dispatch(setRestaurant(restaurant));
-  };
-}
-
-export function requestLogin() {
-  return async (dispatch, getState) => {
-    const { loginFields: { email, password } } = getState();
-
-    const accessToken = await postLogin({ email, password });
-
-    // TODO: 로그인 성공하면 localStorage에 저장한다.
-    saveItem('accessToken', accessToken);
-
-    dispatch(setAccessToken(accessToken));
-  };
-}
-
-export function loadReview({ restaurantId }) {
-  // TODO: fetch restaurant
-  // setReviews
-  return async (dispatch) => {
-    const restaurant = await fetchRestaurant({ restaurantId });
-
-    dispatch(setReviews(restaurant.reviews));
-  };
-}
-
-export function sendReview({ restaurantId }) {
-  return async (dispatch, getState) => {
-    const { accessToken, reviewFields: { score, description } } = getState();
-    await postReview({
-      accessToken, restaurantId, score, description,
-    });
-
-    dispatch(loadReview({ restaurantId }));
-    // dispatch(changeReviewField({ name: 'description', value: '' }));
-    // dispatch(changeReviewField({ name: 'score', value: '' }));
-    dispatch(clearReviewFields());
-  };
-}
