+++ ToGoFurther/src/slice.js
@@ -1,6 +1,17 @@
 // 액션 크리에이터와 리듀서를 동시에 만들어주고 정의까지 한꺼번에 해준다.
 import { createSlice } from '@reduxjs/toolkit';
 
+import {
+  fetchRegions,
+  fetchCategories,
+  fetchRestaurants,
+  fetchRestaurant,
+  postLogin,
+  postReview,
+} from './services/api';
+
+import { saveItem } from './services/storage';
+
 import { equal } from './utils';
 
 const initialReviewFields = {
@@ -128,7 +139,95 @@ const { actions, reducer } = createSlice({
   },
 });
 
-export {
-  actions,
-  reducer,
-};
+export const {
+  setRegions,
+  setCategories,
+  setRestaurant,
+  setRestaurants,
+  selectRegion,
+  selectCategory,
+  changeLoginField,
+  changeReviewField,
+  setAccessToken,
+  logout,
+  clearReviewFields,
+  setReviews,
+} = actions;
+
+export function loadInitialData() {
+  return async (dispatch) => {
+    const regions = await fetchRegions();
+    dispatch(setRegions(regions));
+
+    const categories = await fetchCategories();
+    dispatch(setCategories(categories));
+  };
+}
+
+export function loadRestaurants() {
+  return async (dispatch, getState) => {
+    const {
+      selectedRegion: region,
+      selectedCategory: category,
+    } = getState();
+
+    if (!region || !category) {
+      return;
+    }
+
+    const restaurants = await fetchRestaurants({
+      regionName: region.name,
+      categoryId: category.id,
+    });
+    dispatch(setRestaurants(restaurants));
+  };
+}
+
+export function loadRestaurant({ restaurantId }) {
+  return async (dispatch) => {
+    dispatch(setRestaurant(null));
+
+    const restaurant = await fetchRestaurant({ restaurantId });
+
+    dispatch(setRestaurant(restaurant));
+  };
+}
+
+export function requestLogin() {
+  return async (dispatch, getState) => {
+    const { loginFields: { email, password } } = getState();
+
+    const accessToken = await postLogin({ email, password });
+
+    // TODO: 로그인 성공하면 localStorage에 저장한다.
+    saveItem('accessToken', accessToken);
+
+    dispatch(setAccessToken(accessToken));
+  };
+}
+
+export function loadReview({ restaurantId }) {
+  // TODO: fetch restaurant
+  // setReviews
+  return async (dispatch) => {
+    const restaurant = await fetchRestaurant({ restaurantId });
+
+    dispatch(setReviews(restaurant.reviews));
+  };
+}
+
+export function sendReview({ restaurantId }) {
+  return async (dispatch, getState) => {
+    const { accessToken, reviewFields: { score, description } } = getState();
+    await postReview({
+      accessToken, restaurantId, score, description,
+    });
+
+    dispatch(loadReview({ restaurantId }));
+    // dispatch(changeReviewField({ name: 'description', value: '' }));
+    // dispatch(changeReviewField({ name: 'score', value: '' }));
+    dispatch(clearReviewFields());
+  };
+}
+
+export default reducer;
